refactor ui codebase
